Move actions to self-contained files
