+++ jsconfig.jscon
@@ -7,5 +7,5 @@
       "shared/*": ["server/shared/*"]
     }
   },
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "build"]
 }
